сделал недочеты 1 и 2 домашки
