--- Java15.04.2023/src/main/java/lesson1/dz1/Task3.java
@@ -1,12 +1,0 @@
-package lesson1.dz1;
-
-public class Task3 {
-    public static void main(String[] args) {
-        System.out.println(result(1, 2, 3, 4));
-    }
-
-    public static int result(int a, int b, int c, int d) {
-        return a * (b + (c / d));
-    }
-}
-
