--- Java15.04.2023/src/main/java/lesson1/dz1/Task5.java
@@ -1,15 +1,0 @@
-package lesson1.dz1;
-
-public class Task5 {
-    public static void main(String[] args) {
-        fuu(-2);
-    }
-
-    public static void fuu(int a) {
-        if (a >= 0) {
-            System.out.println("число " + a + " положительное");
-        } else {
-            System.out.println("число " + a + " отрицательное");
-        }
-    }
-}
