--- Java15.04.2023/src/main/java/lesson1/dz1/Task6.java
@@ -1,11 +1,0 @@
-package lesson1.dz1;
-
-public class Task6 {
-    public static void main(String[] args) {
-        System.out.println(fuu(-7));
-    }
-
-    public static boolean fuu(int a) {
-        return a < 0;
-    }
-}
