--- Java15.04.2023/src/main/java/lesson1/dz1/Task7.java
@@ -1,15 +1,0 @@
-package lesson1.dz1;
-
-public class Task7 {
-    public static void main(String[] args) {
-        printName("Bob");
-    }
-
-    public static void printName(String name) {
-        if (name == null || name.isEmpty()) {
-            System.out.println("Имя не задано");
-        } else {
-            System.out.println("Привет, " + name);
-        }
-    }
-}
