--- Java15.04.2023/src/main/java/lesson1/dz2/Task1.java
@@ -1,19 +1,0 @@
-package lesson1.dz2;
-
-import java.util.Arrays;
-
-public class Task1 {
-    public static void main(String[] args) {
-        int[] a = {1, 1, 1, 0, 1, 0, 0,};
-        System.out.println(Arrays.toString(a));
-        for (int i = 0; i < a.length; i++) {
-            if (a[i] == 0) {
-                a[i] = 1;
-            } else if (a[i] == 1) {
-                a[i] = 0;
-            }
-
-        }
-        System.out.println(Arrays.toString(a));
-    }
-}
