--- Java15.04.2023/src/main/java/lesson1/dz2/Task4.java
@@ -1,18 +1,0 @@
-package lesson1.dz2;
-
-import java.util.Arrays;
-
-public class Task4 {
-    public static void main(String[] args) {
-        int[][] arr = new int[7][7];
-        for (int i = 0; i < arr.length; i++) {
-            for (int j = 0, x = arr[i].length - 1; j < arr[i].length; j++, x--) {
-                if (i == j || i == x) arr[i][j] = 1;
-                else arr[i][j] = 0;
-                System.out.print(arr[i][j] + " ");
-            }
-            System.out.print("\n");
-        }
-    }
-}
-
