+++ Java15.04.2023/src/main/java/lesson1/homework1/Task4.java
@@ -1,0 +1,12 @@
+package lesson1.homework1;
+
+public class Task4 {
+    public static void main(String[] args) {
+        System.out.println(fuu(2, 8));
+    }
+
+    public static boolean fuu(int a, int b) {
+        int h = a + b;
+        return h >= 10 && h <= 20;
+    }
+}
