+++ Java15.04.2023/src/main/java/lesson2/homework3/ArrayOfEmployee.java
@@ -1,4 +1,4 @@
-package lesson2.dz3;
+package lesson2.homework3;
 
 public class ArrayOfEmployee {
     public static void main(String[] args) {
